Include normalizer logic in SalesSummary
