+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/dashboard.xml
@@ -22,16 +22,10 @@
         <service id="Sylius\Component\Core\Dashboard\DashboardStatisticsProviderInterface" alias="sylius.dashboard.statistics_provider" />
 
         <service
-            id="Sylius\Component\Core\Dashboard\SalesDataArrayNormalizerInterface"
-            class="Sylius\Component\Core\Dashboard\SalesDataArrayNormalizer"
-        />
-
-        <service
             id="Sylius\Component\Core\Dashboard\SalesDataProviderInterface"
             class="Sylius\Component\Core\Dashboard\SalesDataProvider"
         >
             <argument type="service" id="doctrine.orm.entity_manager" />
-            <argument type="service" id="Sylius\Component\Core\Dashboard\SalesDataArrayNormalizerInterface" />
         </service>
     </services>
 </container>
